Add support for g-prefixed GNU coreutils
Prefixed versions may be used on a system that has a base BSD
coreutils, such as OS X where using MacPorts GNU Coreutils can be
installed, but by default are prefixed with 'g'. Even if PATH is
overriden, the path used contains symbolic links to the g-prefixed
binaries, which means the g-prefixed name is what will appear in the
process list.